+++ lib/common/globals.h
@@ -75,11 +75,11 @@ GLOBALS_API EXTERN int
 GLOBALS_API EXTERN Agsym_t *G_ordering, *G_peripheries, *G_penwidth,
     *G_gradientangle, *G_margin;
 GLOBALS_API EXTERN Agsym_t *N_height, *N_width, *N_shape, *N_color,
-    *N_fillcolor, *N_fontsize, *N_fontname, *N_fontcolor, *N_label, *N_xlabel,
-    *N_nojustify, *N_style, *N_showboxes, *N_sides, *N_peripheries, *N_ordering,
-    *N_orientation, *N_skew, *N_distortion, *N_fixed, *N_imagescale,
-    *N_imagepos, *N_layer, *N_group, *N_comment, *N_vertices, *N_z, *N_penwidth,
-    *N_gradientangle;
+    *N_fillcolor, *N_fontsize, *N_fontname, *N_fontcolor, *N_label,
+    *N_labelwrapwidth, *N_xlabel, *N_nojustify, *N_style, *N_showboxes,
+    *N_sides, *N_peripheries, *N_ordering, *N_orientation, *N_skew,
+    *N_distortion, *N_fixed, *N_imagescale, *N_imagepos, *N_layer, *N_group,
+    *N_comment, *N_vertices, *N_z, *N_penwidth, *N_gradientangle;
 GLOBALS_API EXTERN Agsym_t *E_weight, *E_minlen, *E_color, *E_fillcolor,
     *E_fontsize, *E_fontname, *E_fontcolor, *E_label, *E_xlabel, *E_dir,
     *E_style, *E_decorate, *E_showboxes, *E_arrowsz, *E_constr, *E_layer,